+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Template:
 ### Changed
 - new strategies: peak_load_window
 - plotting time windows with shaded backgrounds
+- changed total power plot: remove individual loads, add sum of batteries, add GC load
 
 ## [1.0.1] - Minor Fixes - 2023-08-04